backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==25.1.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "async-timeout==5.0.1",
    "asyncpg==0.31.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.5",
    "databases==0.9.0",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.133.1",
    "fastapi-cli==0.0.24",
    "fastapi-cloud-cli==0.14.0",
    "fastar==0.8.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "locust>=2.43.3",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "passlib==1.7.4",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.0",
    "pydantic-settings==2.13.1",
    "pygments==2.19.2",
    "pytest>=9.0.2",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "redis==7.2.1",
    "rich==14.3.3",
    "rich-toolkit==0.19.7",
    "rignore==0.7.6",
    "rsa==4.9.1",
    "sentry-sdk==2.53.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sqlalchemy==2.0.47",
    "starlette==0.52.1",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.41.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
]


# ==============================
# 🛠️ Ruff 配置（Lint + Format）
# ==============================
[tool.ruff]
# 启用预览功能（含类型感知 lint、新规则、更智能修复）
preview = true

# 行长（与 formatter 一致）
line-length = 120       # E501 设置 行长

# 忽略的路径（支持 glob）
extend-exclude = [
  ".git",
  "__pycache__",
  "venv",
  ".venv",
  "build",
  "dist",
  "notebooks",        # Jupyter 笔记本通常不 lint
  "zzz_fastapi_backup/*",
  "zzz_sqlalchemy_backup/*",
]

# 自动修复安全问题（CI 中建议开启）
fix = true
unsafe-fixes = false  # 默认不应用可能改变语义的修复

# ==============================
# 🔍 Lint 规则配置
# ==============================
[tool.ruff.lint]
# 启用的规则集（按前缀选择）
# E/W: pycodestyle, F: pyflakes, I: isort,
# UP: pyupgrade, C4: comprehensions, TID: tidy imports,
# ARG: unused args, PTH: pathlib, PL: pylint subset
select = [
  "E", "W", "F",
  "I",
  "UP",
#   "C4",
#   "TID",
#   "ARG",
#   "PTH",
  # "PL",  # pylint 规则较严格，按需启用
]
# 忽略的规则
ignore = [
  "E501",  # 行长交给 formatter 处理
  "D",     # 忽略 docstring 规则（可用 pydoclint 替代）
]

# ==============================
# 🧹 isort 配置（Ruff 内置）
# ==============================
[tool.ruff.lint.isort]
# 已知的一等包（会放在 import 分组顶部）
# known-first-party = ["my_project"]
# 合并 from ... import (a, b) → from ... import a, b
combine-as-imports = true
# 强制按字母排序
force-sort-within-sections = true

# ==============================
# ✨ Formatter 配置（替代 Black）
# ==============================
[tool.ruff.format]
# 使用双引号
quote-style = "double"
